generic graphics format
This adds files with the extension .pzg, which describe
a puzzle on the graphical level, only, as a list of
superimposed components.

For the moment this is not fully implemented, but several
of the one-off types have been converted to .pzg format.
The approach should work fine for a good part of simple
grid-based puzzles, e.g. most sudoku variants.